+++ src/analyze/src/components/pipelines/ConnectorSettings.vue
@@ -1,6 +1,7 @@
 <script>
 import { mapGetters } from 'vuex';
 import InputDateIso8601 from '@/components/generic/InputDateIso8601';
+import TooltipCircle from '@/components/generic/TooltipCircle';
 
 import utils from '@/utils/utils';
 
@@ -8,6 +9,7 @@ export default {
   name: 'ConnectorSettings',
   components: {
     InputDateIso8601,
+    TooltipCircle,
   },
   props: {
     fieldClass: { type: String },
@@ -17,10 +19,23 @@ export default {
       default: () => {},
     },
   },
+  watch: {
+    configSettings: {
+      handler(newVal) {
+        /**
+         * Improve account UX by auto-detecting Account ID via URL
+         */
+
+        const accountInput = newVal.config.account;
+        const parsedAccountId = utils.snowflakeAccountParser(accountInput);
+
+        this.configSettings.config.account = parsedAccountId || newVal.config.account;
+      },
+      deep: true,
+    },
+  },
   computed: {
-    ...mapGetters('configuration', [
-      'getIsConfigSettingValid',
-    ]),
+    ...mapGetters('configuration', ['getIsConfigSettingValid']),
     getCleanedLabel() {
       return value => utils.titleCase(utils.underscoreToSpace(value));
     },
@@ -34,7 +49,8 @@ export default {
       return kind => kind === 'options';
     },
     getIsOfKindTextBased() {
-      return kind => !this.getIsOfKindBoolean(kind) && !this.getIsOfKindDate(kind);
+      return kind =>
+        !this.getIsOfKindBoolean(kind) && !this.getIsOfKindDate(kind);
     },
     getTextBasedInputType() {
       let type = 'text';
@@ -74,30 +90,40 @@ export default {
 <template>
   <div>
     <slot name="top" />
-    <div class="field is-horizontal" v-for='setting in configSettings.settings' :key='setting.name'>
-      <div :class="['field-label', labelClass]">
-        <label class="label">{{ setting.label || getCleanedLabel(setting.name) }}</label>
+    <div
+      class="field is-horizontal"
+      v-for="setting in configSettings.settings"
+      :key="setting.name"
+    >
+      <div :class="['is-flex', 'field-label', labelClass]">
+        <label class="label">{{
+          setting.label || getCleanedLabel(setting.name)
+        }}</label>
+        <TooltipCircle
+          v-if="setting.tooltip"
+          :text="setting.tooltip"
+          class="label-tooltip"
+        />
       </div>
       <div class="field-body">
         <div class="field">
           <div class="control is-expanded">
-
             <!-- Boolean -->
-            <label
-              v-if='getIsOfKindBoolean(setting.kind)'
-              class="checkbox">
+            <label v-if="getIsOfKindBoolean(setting.kind)" class="checkbox">
               <input
                 v-model="configSettings.config[setting.name]"
                 :class="successClass(setting)"
-                type="checkbox">
+                type="checkbox"
+              />
             </label>
 
             <!-- Date -->
             <InputDateIso8601
-              v-else-if='getIsOfKindDate(setting.kind)'
+              v-else-if="getIsOfKindDate(setting.kind)"
               v-model="configSettings.config[setting.name]"
-              :name='setting.name'
-              input-classes='is-small' />
+              :name="setting.name"
+              input-classes="is-small"
+            />
 
             <!-- Dropdown -->
             <div v-else-if="getIsOfKindOptions(setting.kind)" class="select is-small is-fullwidth">
@@ -116,23 +142,18 @@ export default {
 
             <!-- Text / Password / Email -->
             <input
-              v-else-if='getIsOfKindTextBased(setting.kind)'
+              v-else-if="getIsOfKindTextBased(setting.kind)"
               v-model="configSettings.config[setting.name]"
               :class="['input', fieldClass, successClass(setting)]"
               @focus="$event.target.select()"
               :type="getTextBasedInputType(setting)"
-              :placeholder="setting.value || setting.name">
+              :placeholder="setting.value || setting.name"
+            />
           </div>
-          <p
-            v-if="setting.description"
-            class="help is-italic"
-            >
+          <p v-if="setting.description" class="help is-italic">
             {{ setting.description }}
           </p>
-          <p
-            v-if="setting.documentation"
-            class="help"
-            >
+          <p v-if="setting.documentation" class="help">
             <a :href="setting.documentation">More Info.</a>
           </p>
         </div>
@@ -143,4 +164,7 @@ export default {
 </template>
 
 <style lang="scss">
+.label-tooltip  {
+  margin-left: 0.25em;
+}
 </style>