+++ src/analyze/src/utils/utils.js
@@ -135,6 +135,29 @@ export default {
     }
     return singularizeMe;
   },
+  snowflakeAccountParser(account) {
+    const domainFlag = 'snowflakecomputing.com';
+    const domainLocation = account.indexOf(domainFlag);
+    let accountId = '';
+
+    // If the domain exists in user input, assume URL
+    if (domainLocation > -1) {
+      let shortenedUrl = account.slice(
+        0,
+        domainLocation + domainFlag.length,
+      );
+
+      // Clean up URL if http is detected
+      if (shortenedUrl.indexOf('http') > -1) {
+        shortenedUrl = shortenedUrl.slice(shortenedUrl.indexOf('//') + 2);
+      }
+
+      // This could eventually parse data like region if needed
+      accountId = shortenedUrl.split('.')[0];
+    }
+
+    return accountId;
+  },
   titleCase(value) {
     return value
       .replace(